+++ R/cl-def-Ldk.R
@@ -89,7 +89,7 @@ Ldk.Coo <- function(coo, links = NULL, slidings = NULL, fac = coo$fac) {
 #   cat("An Ldk object with: \n")
 #   coo_nb <- length(Ldk)
 #   coo_len <- sapply(Ldk$coo, nrow)
-#   coo_closed <- sapply(Ldk$coo, is_closed)
+#   coo_closed <- sapply(Ldk$coo, coo_is_closed)
 #   cat(" - $coo:", coo_nb, "configurations of landmarks")
 #
 #   # number of coordinates
@@ -120,7 +120,7 @@ print.LdkCoe <- function(x, ...) {
   cat(rep("-", 20), "\n", sep = "")
   coo_nb <- length(Ldk)
   coo_len <- sapply(Ldk$coo, nrow)
-  coo_closed <- sapply(Ldk$coo, is_closed)
+  coo_closed <- sapply(Ldk$coo, coo_is_closed)
   # # number of open outlines cat(' -', coo_nb, 'configurations of landmarks\n') #
   # one random outline eg <- sample(length(Ldk$coo), 1) coo_eg <- Ldk$coo[[eg]]
   # colnames(coo_eg) <- c('x', 'y') cat(' - One random configuration in $coo: '',